ACPI: Comparison of values was fixed by adding a proper casting
Also added "unused" attribute for a parameter in get_table method.

+++ bootloader/ACPI/ExtendedSystemDescriptorTable.cpp
@@ -6,7 +6,7 @@ ExtendedSystemDescriptorTable::ExtendedSystemDescriptorTable(ACPI_XSDT* xsdt) : 
 }
 ACPITableHeader* ExtendedSystemDescriptorTable::get_table(const char* signature)
 {
-    for (int i = 0; i < (get_xsdt()->h.Length - sizeof(get_xsdt()->h)) / sizeof(uint64_t); i++)
+    for (int i = 0; i < static_cast<int>((get_xsdt()->h.Length - sizeof(get_xsdt()->h)) / sizeof(uint64_t)); i++)
     {
         ACPITableHeader* h = (ACPITableHeader *) (uint32_t)get_xsdt()->PointerToOtherSDT[i];
         if (strncmp(h->Signature, signature, 4))
